SNOW-1658905 Extract ConnectorConfigDefinition from the main config class
