+++ src/main/java/com/snowflake/kafka/connector/Utils.java
@@ -404,7 +404,7 @@ public class Utils {
    *
    * @param config input config object
    */
-  static boolean isSnowpipeStreamingIngestion(Map<String, String> config) {
+  public static boolean isSnowpipeStreamingIngestion(Map<String, String> config) {
     return !isSnowpipeIngestion(config);
   }
